Code review

+++ PlantManager/src/main/java/com/ajaskiewicz/PlantManager/service/PlantServiceImpl.java
@@ -1,14 +1,12 @@
 package com.ajaskiewicz.PlantManager.service;
 
 import com.ajaskiewicz.PlantManager.model.Plant;
-import com.ajaskiewicz.PlantManager.model.User;
 import com.ajaskiewicz.PlantManager.repository.PlantRepository;
 import com.ajaskiewicz.PlantManager.repository.UserRepository;
 import com.ajaskiewicz.PlantManager.repository.WateringScheduleRepository;
 import javassist.NotFoundException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -17,17 +15,18 @@ import java.text.SimpleDateFormat;
 import java.util.*;
 
 @Service("plantService")
+@Slf4j
 public class PlantServiceImpl implements PlantService {
 
-    @Autowired
+    private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
+
     private PlantRepository plantRepository;
 
-    @Autowired
     private WateringScheduleRepository wateringScheduleRepository;
 
-    @Autowired
     private UserRepository userRepository;
 
+    @Autowired
     public PlantServiceImpl(PlantRepository plantRepository, WateringScheduleRepository wateringScheduleRepository, UserRepository userRepository) {
         this.plantRepository = plantRepository;
         this.wateringScheduleRepository = wateringScheduleRepository;
@@ -62,37 +61,43 @@ public class PlantServiceImpl implements PlantService {
     }
 
     public Plant createOrUpdatePlant(Plant plant) {
-        System.out.println(plant);
+        log.info("Create request received for: " + plant.toString());
 
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String username = authentication.getName();
-        User queriedUser = userRepository.findByUsername(username);
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        var username = authentication.getName();
+        var queriedUser = userRepository.findByUsername(username);
         plant.setUser(queriedUser);
 
-        if (plant.getId() == null) {
-            plant = plantRepository.save(plant);
-            return plant;
-        } else {
-            plantRepository.save(plant);
-            return plant;
-        }
+
+        plantRepository.save(plant);
+        return plant;
     }
 
     @Override
-    public void delete(int id) {
-        plantRepository.findById(id);
+    public void delete(int id) throws NotFoundException {
+        /*var optionalPlant = plantRepository.findById(id);
+        optionalPlant.orElseThrow(() -> new NotFoundException(String.format("Plant for id %d not found", id)));*/
+
+        var doesExist = plantRepository.existsById(id);
+
+        if(!doesExist) {
+            throw new NotFoundException(String.format("Plant for id %d not found", id));
+        }
+
         plantRepository.deleteById(id);
     }
 
     @Override
     public List<Plant> findPlantsToBeWateredSoon() {
-        List<Plant> allPlants = plantRepository.findAll();
-        List<Plant> plantsToBeWatered = new ArrayList<>();
+        var allPlants = plantRepository.findAll();
+        var plantsToBeWatered = new ArrayList<Plant>();
 
-        for (int index = 0; index < allPlants.size(); index++) {
-            if (findDifferenceInDays(allPlants.get(index).getWateringSchedule().getLast_watered_date(), allPlants.get(index).getWateringSchedule().getWatering_interval()) <= 3
-                    && findDifferenceInDays(allPlants.get(index).getWateringSchedule().getLast_watered_date(), allPlants.get(index).getWateringSchedule().getWatering_interval()) > 0) {
-                plantsToBeWatered.add(allPlants.get(index));
+        for (var i = 0; i < allPlants.size(); i++) {
+            var differenceInDays = findDifferenceInDays(allPlants.get(i).getWateringSchedule().getLastWateredDate(), allPlants.get(i).getWateringSchedule().getWateringInterval());
+            if (differenceInDays <= 3 && differenceInDays > 0) {
+                var plantToBeWatered = allPlants.get(i);
+                plantToBeWatered.setWateringDifferenceInDays(differenceInDays);
+                plantsToBeWatered.add(plantToBeWatered);
             }
         }
 
@@ -100,20 +105,18 @@ public class PlantServiceImpl implements PlantService {
     }
 
     public static long findDifferenceInDays(String lastWateredDate, Integer wateringInterval) {
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-
-        Date date = Calendar.getInstance().getTime();
-        String today = sdf.format(date);
+        Date date = Calendar.getInstance().getTime(); //todo change to LocalDate(in WateringSchedule)
+        String today = DATE_FORMAT.format(date);
 
         long differenceInDays;
         long differenceInTime;
 
         try {
-            Date lwd = sdf.parse(lastWateredDate);
+            Date lwd = DATE_FORMAT.parse(lastWateredDate);
             System.out.println("Last watered date: " + lastWateredDate);
             System.out.println("Watering interval: " + wateringInterval);
 
-            Date t = sdf.parse(today);
+            Date t = DATE_FORMAT.parse(today);
             System.out.println("Today: " + today);
 
             differenceInTime = lwd.getTime() - t.getTime();
@@ -140,7 +143,7 @@ public class PlantServiceImpl implements PlantService {
         List<Integer> differenceInDays = new ArrayList<>();
 
         for (int index = 0; index < plantsToBeWateredSoon.size(); index++) {
-            differenceInDays.add(Math.toIntExact(findDifference(plantsToBeWateredSoon.get(index).getWateringSchedule().getLast_watered_date(), today)));
+            differenceInDays.add(Math.toIntExact(findDifference(plantsToBeWateredSoon.get(index).getWateringSchedule().getlastWateredDate(), today)));
         }
 
         System.out.println(differenceInDays);
@@ -151,16 +154,14 @@ public class PlantServiceImpl implements PlantService {
 
     @Override
     public Plant updateLastWateredDate(Plant plant) {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+        var date = Calendar.getInstance().getTime();
+        var today = DATE_FORMAT.format(date);
 
-        Date date = Calendar.getInstance().getTime();
-        String today = dateFormat.format(date);
+        plant.getWateringSchedule().setLastWateredDate(today);
 
-        plant.getWateringSchedule().setLast_watered_date(today);
-
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String username = authentication.getName();
-        User queriedUser = userRepository.findByUsername(username);
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        var username = authentication.getName();
+        var queriedUser = userRepository.findByUsername(username);
         plant.setUser(queriedUser);
 
         plantRepository.save(plant);
